Om baji/change api names and Tests fixes (#36)
* changed to intuaitie names for school api

* netetr naming

* fixed tests
