+++ __tests__/api/school.test.ts
@@ -1,7 +1,7 @@
-import { POST } from "@/app/api/school/register/route";
-import { GET } from "@/app/api/school/get/route";
-import { PUT } from "@/app/api/school/update/route";
-import { DELETE } from "@/app/api/school/delete/route";
+import { POST } from "@/app/api/school/addSchool/route";
+import { DELETE } from "@/app/api/school/deleteSchool/route";
+import { GET } from "@/app/api/school/listSchool/route";
+import { PUT } from "@/app/api/school/updateSchool/route";
 import SchoolModel from "@/models/schoolModel";
 import { schoolSchema } from "@/models/schoolSchema";
 import { connectDb } from "@/utils/db";
@@ -19,6 +19,7 @@ jest.mock("@/models/schoolModel", () => ({
     findById: jest.fn(),
     findByIdAndUpdate: jest.fn(),
     findByIdAndDelete: jest.fn(),
+    syncIndexes : jest.fn(),
   },
 }));
 
@@ -28,7 +29,14 @@ jest.mock("@/models/schoolSchema", () => ({
   },
 }));
 
-const createMockRequest = (body: any, url = "http://localhost/api/school"): NextRequest => {
+beforeAll(() => {
+  process.env.MONGO_URI = "mongodb://localhost:27017/testdb";
+});
+
+const createMockRequest = (
+  body: any,
+  url = "http://localhost/api/school"
+): NextRequest => {
   return {
     json: () => Promise.resolve(body),
     url,
@@ -39,15 +47,21 @@ describe("API /api/school", () => {
   beforeEach(() => {
     jest.clearAllMocks();
     (connectDb as jest.Mock).mockResolvedValue(undefined);
-  });
-
+  });  
 
   describe("POST /api/school", () => {
     it("should successfully create a new school", async () => {
-      const validSchoolData = { name: "Test School", contact: "1234567890", address: "Test Address" };
+      const validSchoolData = {
+        name: "Test School",
+        contact: "1234567890",
+        address: "Test Address",
+      };
       const mockSchool = { id: "123", save: jest.fn().mockResolvedValue(true) };
 
-      (schoolSchema.safeParse as jest.Mock).mockReturnValue({ success: true, data: validSchoolData });
+      (schoolSchema.safeParse as jest.Mock).mockReturnValue({
+        success: true,
+        data: validSchoolData,
+      });
       (SchoolModel.create as jest.Mock).mockResolvedValue(mockSchool);
 
       const response = await POST(createMockRequest(validSchoolData));
@@ -90,12 +104,12 @@ describe("API /api/school", () => {
           address: "123 Test St",
           extraField: "Ignored",
         };
-  
+
         const mockSchool = {
           save: jest.fn().mockResolvedValue(true),
           id: "mockedId",
         };
-  
+
         (schoolSchema.safeParse as jest.Mock).mockReturnValue({
           success: true,
           data: {
@@ -104,56 +118,53 @@ describe("API /api/school", () => {
             address: "123 Test St",
           },
         });
-  
+
         (SchoolModel.create as jest.Mock).mockResolvedValue(mockSchool);
-  
+
         const response = await POST(createMockRequest(validDataWithExtraField));
         const responseData = await response.json();
-  
+
         expect(SchoolModel.create).toHaveBeenCalledWith({
           name: "Test School",
           contact: "1234567890",
           address: "123 Test St",
         });
-  
+
         expect(responseData).toEqual({
           message: "School Added",
           success: true,
           id: "mockedId",
         });
       });
-  
+
       it("should handle unexpected errors", async () => {
-        (connectDb as jest.Mock).mockImplementation(() => {
-          throw new Error("Unexpected error");
-        });
-  
+        const error = new Error("Database configuration missing");
+        (connectDb as jest.Mock).mockRejectedValue(error);
+
         const validSchoolData = {
           name: "Test School",
           contact: "1234567890",
           address: "123 Test St",
         };
-  
+
         const response = await POST(createMockRequest(validSchoolData));
         const responseData = await response.json();
-  
-        expect(response).toEqual(
-          expect.objectContaining({
-            status: 500,
-          })
-        );
-  
-        expect(responseData).toEqual({
-          message: "Unexpected error",
-          error: expect.any(Object),
-        });
+
+        expect(response.status).toBe(500);
+        expect(responseData).toMatchObject({
+          message: "Database configuration missing",
+          success: false,
+        });        
       });
     });
   });
 
   describe("GET /api/school", () => {
     it("should retrieve all schools", async () => {
-      const mockSchools = [{ id: "1", name: "School 1" }, { id: "2", name: "School 2" }];
+      const mockSchools = [
+        { id: "1", name: "School 1" },
+        { id: "2", name: "School 2" },
+      ];
       (SchoolModel.find as jest.Mock).mockResolvedValue(mockSchools);
 
       const response = await GET(createMockRequest(null));
@@ -161,24 +172,34 @@ describe("API /api/school", () => {
 
       expect(connectDb).toHaveBeenCalled();
       expect(SchoolModel.find).toHaveBeenCalled();
-      expect(responseData).toEqual({ message: "All schools", schools: mockSchools });
+      expect(responseData).toEqual({
+        message: "All schools",
+        schools: mockSchools,
+      });
     });
 
     it("should retrieve a school by ID", async () => {
       const mockSchool = { id: "1", name: "School 1" };
       (SchoolModel.findById as jest.Mock).mockResolvedValue(mockSchool);
 
-      const response = await GET(createMockRequest(null, "http://localhost/api/school?id=1"));
+      const response = await GET(
+        createMockRequest(null, "http://localhost/api/school?id=1")
+      );
       const responseData = await response.json();
 
       expect(SchoolModel.findById).toHaveBeenCalledWith("1");
-      expect(responseData).toEqual({ message: "School found", school: mockSchool });
+      expect(responseData).toEqual({
+        message: "School found",
+        school: mockSchool,
+      });
     });
 
     it("should return 404 if school not found", async () => {
       (SchoolModel.findById as jest.Mock).mockResolvedValue(null);
 
-      const response = await GET(createMockRequest(null, "http://localhost/api/school?id=1"));
+      const response = await GET(
+        createMockRequest(null, "http://localhost/api/school?id=1")
+      );
       const responseData = await response.json();
 
       expect(response.status).toBe(404);
@@ -188,16 +209,30 @@ describe("API /api/school", () => {
 
   describe("PUT /api/school", () => {
     it("should successfully update a school", async () => {
-      const validData = { id: "1", name: "Updated School", contact: "1234567890", address: "New Address" };
+      const validData = {
+        id: "1",
+        name: "Updated School",
+        contact: "1234567890",
+        address: "New Address",
+      };
       const updatedSchool = { ...validData };
 
-      (schoolSchema.safeParse as jest.Mock).mockReturnValue({ success: true, data: validData });
-      (SchoolModel.findByIdAndUpdate as jest.Mock).mockResolvedValue(updatedSchool);
+      (schoolSchema.safeParse as jest.Mock).mockReturnValue({
+        success: true,
+        data: validData,
+      });
+      (SchoolModel.findByIdAndUpdate as jest.Mock).mockResolvedValue(
+        updatedSchool
+      );
 
       const response = await PUT(createMockRequest(validData));
       const responseData = await response.json();
 
-      expect(SchoolModel.findByIdAndUpdate).toHaveBeenCalledWith("1", validData, { new: true });
+      expect(SchoolModel.findByIdAndUpdate).toHaveBeenCalledWith(
+        "1",
+        validData,
+        { new: true }
+      );
       expect(responseData).toEqual({
         message: "School Updated",
         success: true,
@@ -210,13 +245,17 @@ describe("API /api/school", () => {
       const responseData = await response.json();
 
       expect(response.status).toBe(400);
-      expect(responseData).toEqual({ message: "Request body and ID are required" });
+      expect(responseData).toEqual({
+        message: "Request body and ID are required",
+      });
     });
 
     it("should return 404 if school not found", async () => {
       (SchoolModel.findByIdAndUpdate as jest.Mock).mockResolvedValue(null);
 
-      const response = await PUT(createMockRequest({ id: "1", name: "Updated School" }));
+      const response = await PUT(
+        createMockRequest({ id: "1", name: "Updated School" })
+      );
       const responseData = await response.json();
 
       expect(response.status).toBe(404);
